Calculate Standard Error and Save Variables
Certain parameters, like the standard deviation, and number of rows,
were saved as variables, to make writing code later on easier.

The Standard Error was calculated using the standard deviation and
number of samples for the control. The control was chosen as the base,
since Cognitive Behaviour Therapy is the older and more common form, and
the one we're comparing the VR Therapy to.